+++ src/Network/Network/ChangeLog.md
@@ -35,6 +35,7 @@
 * New Generations and SKUs for VirtualNetworkGateways
   - Introduce new Generations for VirtualNetworkGateways.
   - Introduce new high throughput SKUs for VirtualNetworkGateways.
+* New cmdlets to support the azure firewall policy
 
 ## Version 1.14.0
 * Fix incorrect example in `New-AzApplicationGateway` reference documentation 
@@ -69,7 +70,6 @@
 * Adjusted validation logic for Ipv6 IP Prefix to check for correct IPv6 prefix length.
 * Updated Get-AzVirtualNetworkSubnetConfig: Added parameter set to get by subnet resource id.
 * Updated description of Location parameter for AzNetworkServiceTag
-* New commandlets to support the azure firewall policy
 
 ## Version 1.12.0
 * Add support for private endpoint and private link service